Refactor require
